Add getCookieValue method to handle cookie retrieval
Introduced the getCookieValue method in request.ts to streamline the extraction of cookie values from request headers. This method streamlines code reuse, centralizing in one location a common task that may be required in different parts of the application. This function retrieves the value of a cookie given its name or returns null if it doesn't exist.