Add urlToCapitalizedWords function to string utility
The `urlToCapitalizedWords` function has been added to the string utility. This function takes a URL as an input, extracts all words from the pathname and returns them as an array of capitalized words. This is useful for extracting and formatting words from URLs for usage in other parts of the application, such as generating readable text from URLs.