Change `BetterNumber` class's `number` type to `unknown`
The `number` field of `BetterNumberProperties` in `src/better-number.ts` is changed from `number | bigint` to `unknown`. This change allows the constructor of `BetterNumber` to assert the input `number` as a number. It resolves issues where invalid strings might get passed as number, causing unexpected behaviors. Now, an invalid input will be coerced to `undefined`. The purpose behind this change is to enhance the robustness by ensuring better input validation.